+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - [FIX] Objc attributes interop for KMP. See [#1947][]
+- [FIX] Inject backtrace reporter into Logs feature. See [#1948][]
 
 # 2.14.0 / 04-07-2024
 
@@ -711,6 +712,7 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#1934]: https://github.com/DataDog/dd-sdk-ios/pull/1934
 [#1938]: https://github.com/DataDog/dd-sdk-ios/pull/1938
 [#1947]: https://github.com/DataDog/dd-sdk-ios/pull/1947
+[#1948]: https://github.com/DataDog/dd-sdk-ios/pull/1948
 [@00fa9a]: https://github.com/00FA9A
 [@britton-earnin]: https://github.com/Britton-Earnin
 [@hengyu]: https://github.com/Hengyu